renamed log.py to logging.py to avoid conflict
